Fix handling of forced python version.
We run into problems if we configure with a python version different from that
found by the default find_dependency call for Python.

We generate a cmake fragment with the lines

```
if(@REQUIRE_PYTHON_VERSION@)
  find_dependency(Python @REQUIRE_PYTHON_VERSION@ COMPONENTS Interpreter)
else()
  find_dependency(Python COMPONENTS Interpreter)
endif()
```

The if condition expands to something like

```
if(3.12.9)
```

cmake tries to interpret the argument here as a floating-point number,
but fails to convert it, and (silently!) treats the condition as false.
Need to do this test explictly as a string comparison instead.